+++ web-client/integration-tests/journey/admissionsClerkEditsPetitionerEmail.js
@@ -1,0 +1,67 @@
+import { SERVICE_INDICATOR_TYPES } from '../../../shared/src/business/entities/EntityConstants';
+import {
+  contactPrimaryFromState,
+  refreshElasticsearchIndex,
+} from '../helpers';
+
+export const admissionsClerkEditsPetitionerEmail = (
+  cerebralTest,
+  emailToAdd,
+) => {
+  return it('admissions clerk adds petitioner email with existing cognito account to case', async () => {
+    await refreshElasticsearchIndex();
+
+    let contactPrimary = contactPrimaryFromState(cerebralTest);
+
+    await cerebralTest.runSequence(
+      'gotoEditPetitionerInformationInternalSequence',
+      {
+        contactId: contactPrimary.contactId,
+        docketNumber: cerebralTest.docketNumber,
+      },
+    );
+
+    expect(cerebralTest.getState('currentPage')).toEqual(
+      'EditPetitionerInformationInternal',
+    );
+    expect(cerebralTest.getState('form.updatedEmail')).toBeUndefined();
+    expect(cerebralTest.getState('form.confirmEmail')).toBeUndefined();
+
+    await cerebralTest.runSequence('updateFormValueSequence', {
+      key: 'contact.updatedEmail',
+      value: emailToAdd,
+    });
+
+    await cerebralTest.runSequence('updateFormValueSequence', {
+      key: 'contact.confirmEmail',
+      value: emailToAdd,
+    });
+
+    await cerebralTest.runSequence('submitEditPetitionerSequence');
+
+    expect(cerebralTest.getState('validationErrors')).toEqual({});
+
+    expect(cerebralTest.getState('modal.showModal')).toBe(
+      'MatchingEmailFoundModal',
+    );
+    expect(cerebralTest.getState('currentPage')).toEqual(
+      'EditPetitionerInformationInternal',
+    );
+
+    await cerebralTest.runSequence(
+      'submitUpdatePetitionerInformationFromModalSequence',
+    );
+
+    expect(cerebralTest.getState('modal.showModal')).toBeUndefined();
+    expect(cerebralTest.getState('currentPage')).toEqual('CaseDetailInternal');
+
+    contactPrimary = contactPrimaryFromState(cerebralTest);
+
+    expect(contactPrimary.email).toEqual(emailToAdd);
+    expect(contactPrimary.serviceIndicator).toEqual(
+      SERVICE_INDICATOR_TYPES.SI_ELECTRONIC,
+    );
+
+    await refreshElasticsearchIndex();
+  });
+};
